requirements.txt
pymongo
dnspython
certifi
pandas
numpy
scikit-learn
xgboost
streamlit
python-dotenv
requests
matplotlib
seaborn
joblib
shap